Module functions

Domains.Domain — Type

Domain definitions

Domains.Domain — Method

Domain(d::Dict)

Constructor for domains from Dictionaries

Domains.Domain — Method

Domain(name::String)

Constructor for Domain

Base.in — Method

in(lonlat, d; with_ezone = false) 
lonlat in d 
lonlat ∈ d

Returns true if lonlat is inside Domain d

Domains.Plcc — Method

Plcc(d)

Returns a Lambert Conformal Conic Proj4.Projection based on domain definitions in d

Domains.Plonlat — Method

Plonlat()

Returns longlat Proj4.Projection

Domains.get_lcc_val — Method

get_lcc_val(d)

Returns a named tuple with boundary values of the domain in projected coordinates

Domains.getgridpoints — Method

getgridpoints(d; gsize=d.GSIZE)

Returns an array with the lonlat coordinates of the grid points

Domains.lcc2lonlat — Method

lcc2lonlat(d,xy)

Returns lonlat coordinates for Lambert Conformal Conic projection coordinates xy using domain definitions from d

Domains.lonlat2lcc — Method

lonlat2lcc(d,lonlat)

Returns Lambert Conformal Conic projection coordinates for lonlat using domain definition from d